# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for GroceryTrak.",
        "title": "GroceryTrak API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "GroceryTrak",
            "email": "grocerytrak@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/recipe": {
            "post": {
                "description": "Creates a new recipe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipe"
                ],
                "summary": "Create a recipe",
                "parameters": [
                    {
                        "description": "Recipe Data",
                        "name": "recipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.RecipeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.RecipeResponse"
                        }
                    },
                    "default": {
                        "description": "Standard Error Responses",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/recipe/search": {
            "get": {
                "description": "Searches for recipes by title, ingredients, or diet type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipe"
                ],
                "summary": "Search recipes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title of recipe",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated ingredient IDs",
                        "name": "ingredients",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Diet type",
                        "name": "diet",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.RecipesResponse"
                        }
                    },
                    "default": {
                        "description": "Standard Error Responses",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/recipe/{id}": {
            "get": {
                "description": "Retrieves a recipe by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipe"
                ],
                "summary": "Get a recipe",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recipe ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.RecipeResponse"
                        }
                    },
                    "default": {
                        "description": "Standard Error Responses",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing recipe by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipe"
                ],
                "summary": "Update a recipe",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recipe ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Recipe Data",
                        "name": "recipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.RecipeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.RecipeResponse"
                        }
                    },
                    "default": {
                        "description": "Standard Error Responses",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a recipe by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipe"
                ],
                "summary": "Delete a recipe",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recipe ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "default": {
                        "description": "Standard Error Responses",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "An error occurred"
                }
            }
        },
...
        "dtos.ItemResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image": {
                    "type": "string",
                    "example": "milk.jpg"
                },
                "name": {
                    "type": "string",
                    "example": "Milk"
                },
                "spoonacular_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "dtos.RecipeItemRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 2.5
                },
                "item_id": {
                    "type": "integer",
                    "example": 456
                },
                "unit": {
                    "type": "string",
                    "example": "cups"
                }
            }
        },
        "dtos.RecipeItemResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 2.5
                },
                "item": {
                    "$ref": "#/definitions/dtos.ItemResponse"
                },
                "unit": {
                    "type": "string",
                    "example": "cups"
                }
            }
        },
        "dtos.RecipeRequest": {
            "type": "object",
            "properties": {
                "cooking_time": {
                    "type": "integer",
                    "example": 20
                },
                "image": {
                    "type": "string",
                    "example": "https://example.com/spaghetti.jpg"
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.RecipeItemRequest"
                    }
                },
                "kcal": {
                    "type": "number",
                    "example": 450.5
                },
                "prep_time": {
                    "type": "integer",
                    "example": 10
                },
                "ready_time": {
                    "type": "integer",
                    "example": 30
                },
                "title": {
                    "type": "string",
                    "example": "Spaghetti Carbonara"
                },
                "vegan": {
                    "type": "boolean",
                    "example": false
                },
                "vegetarian": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "dtos.RecipeResponse": {
            "type": "object",
            "properties": {
                "cooking_time": {
                    "type": "integer",
                    "example": 20
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image": {
                    "type": "string",
                    "example": "https://example.com/spaghetti.jpg"
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.RecipeItemResponse"
                    }
                },
                "kcal": {
                    "type": "number",
                    "example": 450.5
                },
                "prep_time": {
                    "type": "integer",
                    "example": 10
                },
                "ready_time": {
                    "type": "integer",
                    "example": 30
                },
                "title": {
                    "type": "string",
                    "example": "Spaghetti Carbonara"
                },
                "vegan": {
                    "type": "boolean",
                    "example": false
                },
                "vegetarian": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "dtos.RecipesResponse": {
            "type": "object",
            "properties": {
                "recipes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.RecipeResponse"
                    }
                }
            }
        },
...
    }
}